feat: implement metrics management with cubit and database integration

+++ lib/presentation/dashboard/widgets/daily_goal_card.dart
@@ -11,7 +11,7 @@ class DailyGoalCard extends StatelessWidget {
   });
 
   final int currentTimeInvested;
-  final int dailyGoalInMinutes;
+  final double dailyGoalInMinutes;
   final VoidCallback onStart;
 
   @override
